# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API for Users",
    "version": "1.0",
    "description": "Api untuk aplikasi users",
    "termsOfService": "https://github.com/zaenal-abidin-syah",
    "contact": {
      "name": "Zaenal Abidin Syah",
      "email": "aszaenal618@gmail.com",
      "url": "https://github.com/zaenal-abidin-syah"
    },
    "license": {
      "name": "APACHE 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "https://api.zaens.my.id",
      "description": "Users Restfull Api server"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get All Users",
        "description": "Get All Users by default",
        "parameters": [],
        "responses": {
          "200": {
            "description": "success get all users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArrayUsers"
                },
                "examples": {
                  "success": {
                    "$ref": "#/components/examples/successGetAllUsers"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create New Users",
        "description": "Create New Users to database",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUsers"
              },
              "examples": {
                "contoh1": {
                  "$ref": "#/components/examples/contoh1"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success create Users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                },
                "examples": {
                  "contoh1": {
                    "description": "Success Create Users",
                    "value": {
                      "message": "create User success"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed create Users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                },
                "examples": {
                  "contoh1": {
                    "description": "Failed Create Users",
                    "value": {
                      "message": "Some fields are missing!!"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get User by User Id",
        "description": "Get user by user id",
        "parameters": [
          {
            "$ref": "#/components/parameters/userId"
          }
        ],
        "responses": {
          "200": {
            "description": "success get user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                },
                "examples": {
                  "success": {
                    "$ref": "#/components/examples/successGetUser"
                  }
                }
              }
            }
          },
          "404": {
            "description": "failed get users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "failed": {
                    "description": "failed get user",
                    "value": {
                      "message": "user not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update full User by User Id",
        "description": "Update full user by user id",
        "parameters": [
          {
            "$ref": "#/components/parameters/userId"
          }
        ],
        "responses": {
          "200": {
            "description": "success update full users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "success": {
                    "description": "success update user",
                    "value": {
                      "message": "update user success"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "some field missing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "missing": {
                    "description": "some field missing",
                    "value": {
                      "message": "Some fields are missing!!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user id not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "idNotFound": {
                    "description": "user id not found",
                    "value": {
                      "message": "user not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Update Separately User by User Id",
        "description": "Update Separately user by user id",
        "parameters": [
          {
            "$ref": "#/components/parameters/UsersId"
          }
        ],
        "responses": {
          "200": {
            "description": "success update Separately users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "success": {
                    "description": "success update user",
                    "value": {
                      "message": "update user success"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user id not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "idNotFound": {
                    "description": "user id not found",
                    "value": {
                      "message": "user not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete User by User Id",
        "description": "Delete user by user id",
        "parameters": [
          {
            "$ref": "#/components/parameters/userId"
          }
        ],
        "responses": {
          "200": {
            "description": "success Delete users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "success": {
                    "description": "success delete user",
                    "value": {
                      "message": "delete user success"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user id not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "idNotFound": {
                    "description": "user id not found",
                    "value": {
                      "message": "user not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "userId": {
        "name": "userId",
        "in": "path",
        "required": true,
        "description": "User id for get user",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 100,
          "example": "1"
        },
        "examples": {
          "contoh1": {
            "description": "sample todolist for contoh1",
            "value": "1"
          },
          "contoh2": {
            "description": "sample todolist for contoh2",
            "value": "2"
          }
        }
      }
    },
    "schemas": {
      "Users": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "number_phone": {
            "type": "string"
          }
        }
      },
      "ArrayUsers": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Users"
        }
      },
      "CreateUsers": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "required": true,
            "minLength": 4
          },
          "username": {
            "type": "string",
            "required": true,
            "minLength": 4
          },
          "email": {
            "type": "string",
            "required": true,
            "minLength": 4
          },
          "number_phone": {
            "type": "string",
            "required": true,
            "minLength": 4
          }
        }
      }
    },
    "examples": {
      "successGetAllUsers": {
        "description": "success get all users",
        "value": [
          {
            "id": 99,
            "name": "contoh1",
            "username": "contoh1",
            "email": "contoh1@email.com",
            "phone_number": "083265545655"
          },
          {
            "id": 100,
            "name": "contoh2",
            "username": "contoh2",
            "email": "contoh2@email.com",
            "phone_number": "083475348435"
          }
        ]
      },
      "successGetUser": {
        "description": "success get user",
        "value": {
          "id": 99,
          "name": "contoh1",
          "username": "contoh1",
          "email": "contoh1@email.com",
          "phone_number": "083265545655"
        }
      },
      "contoh1": {
        "description": "create users contoh1",
        "value": {
          "name": "name1",
          "username": "username1",
          "email": "username1@email.com",
          "phone_number": "088946283732"
        }
      }
    }
  }
}